refactor: RealisticAlgorithm을 algorithm 패키지로 이동
HoldOutAlgorithm이 goal.service.algorithm에 자리를 잡아, 추천 알고리즘
구현체가 모이는 곳이 그쪽으로 정해졌다. 같은 성격의 클래스가 두 패키지에
흩어져 있을 이유가 없어 맞췄다.

테스트도 같은 경로로 옮겼다.
