+++ src/main/java/com/team/independence/goal/service/algorithm/RealisticAlgorithm.java
@@ -1,4 +1,4 @@
-package com.team.independence.goal.service;
+package com.team.independence.goal.service.algorithm;
 
 import com.team.independence.asset.dto.summary.AssetNetWorthBreakdown;
 import com.team.independence.asset.dto.summary.AssetSummaryResponse;
@@ -7,6 +7,9 @@ import com.team.independence.goal.dto.AlgorithmType;
 import com.team.independence.goal.dto.GoalRecommendationRequest;
 import com.team.independence.goal.dto.GoalRecommendationResponse;
 import com.team.independence.goal.dto.LoanPlans;
+import com.team.independence.goal.service.GoalService;
+import com.team.independence.goal.service.LoanPlanCalculator;
+import com.team.independence.goal.service.RecommendationAlgorithm;
 import com.team.independence.property.domain.DealType;
 import com.team.independence.property.domain.HousingType;
 import com.team.independence.property.dto.RentMedianRequest;
