+++ src/test/java/com/team/independence/goal/service/algorithm/RealisticAlgorithmTest.java
@@ -1,4 +1,4 @@
-package com.team.independence.goal.service;
+package com.team.independence.goal.service.algorithm;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
@@ -15,6 +15,8 @@ import com.team.independence.goal.dto.AlgorithmType;
 import com.team.independence.goal.dto.GoalRecommendationRequest;
 import com.team.independence.goal.dto.GoalRecommendationResponse.RecommendationItem;
 import com.team.independence.goal.dto.LoanPlans;
+import com.team.independence.goal.service.GoalService;
+import com.team.independence.goal.service.LoanPlanCalculator;
 import com.team.independence.property.domain.DealType;
 import com.team.independence.property.domain.HousingType;
 import com.team.independence.property.dto.RentMedianRequest;
